instruct annex to commit text files straight to git
There is no configuration for any files to go to git directly,
that seems to make it tricky to manage text files here and they
would tend to auto-migrate into git-annex, e.g. as unlocked git annexed
files.  With such setting the management of this repo would be made easier